# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    name: Librarium
    url: https://github.com/fireball1725/librarium
  description: Self-hosted library tracking API. All protected endpoints require a
    Bearer JWT in the Authorization header.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  termsOfService: http://localhost/
  title: Librarium API
  version: 26.0.0-dev
paths:
...
  /libraries/{library_id}/books/{book_id}/cover:
    delete:
      description: Removes the cover image from a book.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete book cover
      tags:
      - covers
    get:
      description: Streams the primary cover image for a book.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Serve book cover image
      tags:
      - covers
    put:
      consumes:
      - multipart/form-data
      description: Accepts a multipart form upload and stores the image as the book's
        cover.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Cover image file
        in: formData
        name: cover
        required: true
        type: file
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload book cover
      tags:
      - covers
  /libraries/{library_id}/books/{book_id}/cover/fetch:
    post:
      consumes:
      - application/json
      description: Downloads a cover image from the given URL and stores it as the
        book's cover.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Cover URL
        in: body
        name: body
        required: true
        schema:
          properties:
            url:
              type: string
          type: object
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Fetch book cover from URL
      tags:
      - covers
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"